password_strength_checker/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User


class UserSignupSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)
    password2 = serializers.CharField(write_only=True)
    class Meta:
        model = User
        fields = ['first_name', 'last_name', 'username', 'email', 'password', 'password2']

    def validate(self, attrs):
        if attrs.get('password') != attrs.get('password2'):
            raise serializers.ValidationError({"password" : "Passwords do not match"}) 
        return attrs

    def create(self, validated_data):
        validated_data.pop("password2")
        user = User.objects.create_user(**validated_data)
        return user



class PasswordStrengthSerializer(serializers.Serializer):
    password = serializers.CharField(write_only=True, max_length=128)
